requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0

# LangGraph for workflow management
langgraph==0.2.16
# langgraph depends on langchain-core<0.3, so we do not pin langchain-core here to avoid conflicts

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging
loguru==0.7.2

# Data Processing
pandas==2.1.3
openpyxl==3.1.2
python-dateutil==2.8.2

# HTTP Clients
httpx==0.25.2
requests==2.31.0

# Web Scraping (for LinkedIn)
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Google Sheets Integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# AI/ML (Optional)
openai==1.6.1
anthropic==0.7.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
